LayoutTests/accessibility/mac/live-regions/live-region-relevant-text-non-atomic-expected.txt

This tests that aria-relevant=text properly announces when setting textContent on a live region.

Received announcement request:
AnnouncementKey: New Content{
    AXLanguage = en;
}
AXPriorityKey: 10
AXAnnouncementIsLiveRegionKey: true


PASS successfullyParsed is true

TEST COMPLETE
New Content
